# src/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SpotiBye API
  description: Cloudflare Workers backend for Spotify playlist analysis and export functionality
  version: 1.0.0
  contact:
    name: SpotiBye Team
    email: support@spotibye.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://spotibye-api.workers.dev
    description: Production server
  - url: https://spotibye-api-dev.workers.dev
    description: Development server
  - url: http://localhost:8787
    description: Local development server
paths:
...
  /auth/spotify/login:
    post:
      tags:
        - Authentication
      summary: Initiate Spotify OAuth flow
      description: Starts the OAuth authentication process with Spotify
      operationId: spotifyLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpotifyLoginRequest'
            examples:
              standard:
                summary: Standard login request
                value:
                  redirect_uri: "http://localhost:3000/callback"
                  state: "random-state-string"
      responses:
        '200':
          description: OAuth URL generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpotifyLoginResponse'
              examples:
                success:
                  summary: Successful login initiation
                  value:
                    data:
                      auth_url: "https://accounts.spotify.com/authorize?response_type=code&client_id=..."
                      state: "random-state-string"
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/spotify/callback:
    get:
      tags:
        - Authentication
      summary: Handle Spotify OAuth callback
      description: Processes the OAuth callback from Spotify and returns JWT tokens
      operationId: spotifyCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code from Spotify
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: State parameter for CSRF protection
        - name: error
          in: query
          required: false
          schema:
            type: string
          description: Error parameter if OAuth failed
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpotifyCallbackResponse'
              examples:
                success:
                  summary: Successful authentication
                  value:
                    data:
                      access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      refresh_token: "refresh-token-string"
                      expires_in: 3600
                      user:
                        id: "user-123"
                        display_name: "Test User"
                        email: "test@example.com"
        '400':
          description: Invalid callback parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /analysis/playlist/{id}:
    post:
      tags:
        - Analysis
      summary: Analyze playlist
      description: Performs detailed analysis on a playlist including audio features and statistics
      operationId: analyzePlaylist
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Spotify playlist ID
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisRequest'
            examples:
              standard:
                summary: Standard analysis request
                value:
                  include_audio_features: true
                  include_recommendations: false
      responses:
        '200':
          description: Analysis completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
              examples:
                success:
                  summary: Successful analysis
                  value:
                    data:
                      analysis:
                        playlist_id: "playlist-123"
                        total_tracks: 100
                        duration_minutes: 360
                        average_bpm: 128.5
                        energy_score: 0.75
                        danceability: 0.82
                        valence: 0.68
                        acousticness: 0.25
                        instrumentalness: 0.15
                        genres:
                          - "Pop"
                          - "Electronic"
                          - "Rock"
                        top_artists:
                          - name: "Artist One"
                            count: 8
                          - name: "Artist Two"
                            count: 5
                        audio_features:
                          bpm_distribution:
                            "60-80": 10
                            "80-100": 20
                            "100-120": 30
                            "120-140": 25
                            "140-160": 10
                            "160+": 5
                        recommendations:
                          similar_playlists:
                            - id: "similar-123"
                              name: "Similar Vibes"
                              similarity_score: 0.85
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Playlist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Analysis failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Spotify OAuth callback
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
      required:
        - error
    SpotifyLoginRequest:
      type: object
      properties:
        redirect_uri:
          type: string
          format: uri
          description: URL to redirect to after OAuth completion
        state:
          type: string
          description: State parameter for CSRF protection
      required:
        - redirect_uri
    SpotifyLoginResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            auth_url:
              type: string
              format: uri
              description: Spotify authorization URL
            state:
              type: string
              description: State parameter
      required:
        - data
    SpotifyCallbackResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            access_token:
              type: string
              description: JWT access token
            refresh_token:
              type: string
              description: JWT refresh token
            expires_in:
              type: integer
              description: Token expiration time in seconds
            user:
              type: object
              properties:
                id:
                  type: string
                  description: Spotify user ID
                display_name:
                  type: string
                  description: User's display name
                email:
                  type: string
                  format: email
                  description: User's email
              required:
                - id
                - display_name
      required:
        - data
...
    AnalysisRequest:
      type: object
      properties:
        include_audio_features:
          type: boolean
          default: true
          description: Whether to include audio features analysis
        include_recommendations:
          type: boolean
          default: false
          description: Whether to include playlist recommendations
    AnalysisResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            analysis:
              type: object
              properties:
                playlist_id:
                  type: string
                  description: Analyzed playlist ID
                total_tracks:
                  type: integer
                  description: Total number of tracks
                duration_minutes:
                  type: number
                  description: Total playlist duration in minutes
                average_bpm:
                  type: number
                  description: Average BPM across tracks
                energy_score:
                  type: number
                  description: Average energy score (0-1)
                danceability:
                  type: number
                  description: Average danceability score (0-1)
                valence:
                  type: number
                  description: Average valence score (0-1)
                acousticness:
                  type: number
                  description: Average acousticness score (0-1)
                instrumentalness:
                  type: number
                  description: Average instrumentalness score (0-1)
                genres:
                  type: array
                  items:
                    type: string
                  description: Detected genres in playlist
                top_artists:
                  type: array
                  items:
                    $ref: '#/components/schemas/TopArtist'
                  description: Most frequent artists
                audio_features:
                  type: object
                  properties:
                    bpm_distribution:
                      type: object
                      description: BPM distribution across ranges
                recommendations:
                  type: object
                  properties:
                    similar_playlists:
                      type: array
                      items:
                        $ref: '#/components/schemas/SimilarPlaylist'
                  description: Playlist recommendations
              required:
                - playlist_id
                - total_tracks
                - duration_minutes
      required:
        - data
    TopArtist:
      type: object
      properties:
        name:
          type: string
          description: Artist name
        count:
          type: integer
          description: Number of tracks by this artist
      required:
        - name
        - count
    SimilarPlaylist:
      type: object
      properties:
        id:
          type: string
          description: Playlist ID
        name:
          type: string
          description: Playlist name
        similarity_score:
          type: number
          description: Similarity score (0-1)
      required:
        - id
        - name
        - similarity_score
...